refactor: consolidate all card config into CardConfig compound component
