+++ src/components/configurations/index.ts
@@ -1,2 +1,1 @@
-export { CardConfig } from './CardConfig'
 export { cardConfigurations } from './cardConfigurations'
